test_requirements.txt
# Test-only dependencies are unpinned.
#
requests[security]
pip
coverage>=3.6
mock>=1.0.1
nose>=1.3.1
flake8
-e bzr+http://bazaar.launchpad.net/~bigdata-dev/charm-helpers/framework/#egg=charmhelpers
#
# Specify precise versions of runtime dependencies where possible.
PyYAML==3.10       # precise
path.py>=7.0
jujuresources>=0.2.5
